refactored signin/signup & renamed signin->login and src->screens
