+++ client/screens/Navigation/HomeStack.js
@@ -2,8 +2,9 @@ import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 //import { createDrawerNavigator } from '@react-navigation/drawer';
 
-import { SignIn } from "../SignIn/SignIn";
-import { SignUp } from "../SignIn/SignUp";
+// import { SignIn } from "../Authentication/SignIn";
+// import { SignUp } from "../Authentication/SignUp";
+import AuthScreen from "../Authentication/AuthScreen";
 import { Profile } from "../Profile/Profile";
 import { Toggle } from "../Toggle/Toggle";
 import { Chat } from "../Chat/Chat";
@@ -24,8 +25,7 @@ function DrawerNavigator() {
 export default function Navigator() {
   return (
     <Stack.Navigator screenOptions={{ headerShown: false }}>
-      <Stack.Screen name="Sign In" component={SignIn} />
-      <Stack.Screen name="Sign Up" component={SignUp} />
+      <Stack.Screen name="AuthScreen" component={AuthScreen} />
       <Stack.Screen name="Toggle" component={Toggle} />
       <Stack.Screen name="Profile" component={Profile} />
       <Stack.Screen name="Chat" component={Chat} />
